frontend: Port AWS SSH selection to new form stuff
- Move Selector from AWS VPC to UI.
- Use Selector on AWS SSH page.
